Fix service external ips update
The exit node ip was not removed from the service externaIPs property.
Now we remove only the ip from the exit node leaving any other ip that
might be added to the service by others untouched.

When the tunnel is created, the exit nodes ip is added to the list
instead of overriding all externalIPs

Signed-off-by: Juan Ignacio Donoso <jidonoso@gmail.com>