Fix deployment gets removed after tunnel
If a managed services was removed, the tunnel was removed too but the
deployment with inlets client got orphaned. To fix this we set the
tunnel object to be the owner of the deployment instead of the
deployment template, this way kubernetes GC automatically removes the
deployement when the tunnel is removed.

Signed-off-by: Juan Ignacio Donoso <jidonoso@gmail.com>