Fix bug in ownerReference
The ownerReference had the wrong type, kind and group - that of
the tunnel and not that of the service which caused the tunnel
to be created. Tested e2e and shown to be fixed, the problem
was observed when rebooting a machine running kubeadm or an RPi
running k3s.

Thanks to @stefanprodan for helping me to find this and to
@utsavanand2 for confirming the issue.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>