Fix issue with CRD required fields
The ClientDeploymentRef field was marked by the new generated
CRD YAML file as required. This resulted in a bug where the
operator simply would not work anymore due to client_deployment
being null (the starting state) and that being rejected by
validation.

This patch fixes the issue by decorating the field. A Tunnel
YAML was applied successfully after the patched YAML was
applied.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>