Fix issue with status sub-resource
After moving to a status sub-resource in the generated CRD
the status had to be updated via UpdateStatus() on the
Kubernetes clientset, not via Update() as we had it.

The CRD is also made so that its status fields are optional
so the status can be set whilst we have a hostID, but no
hostIP yet.

Requires a new release and helm chart version for the users to
receive this fix.

A big thank you to @munnerz for the hints on this.

Fixes: #84

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>