Migrate to inlets PRO 0.7.0
Migrates to inlets PRO 0.7.0 which makes breaking changes to
flags used changing the way the server is provisioned.

The client for inlets-pro is also updated to use the new
flags --url and --ports and --upstream.

De-duplicates the userdata code for VMs, and brings that in
from the provsion package in inletsctl.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>